feat: Introduce defjvp function for JAX primitives and refactor related imports
